fix: dto ratio타입 entity와 동일하게 double로 변경

+++ pickle-common/src/main/java/com/example/pickle_common/strategy/dto/CreateStrategyRequestDto.java
@@ -34,7 +34,7 @@ public class CreateStrategyRequestDto {
         private String category;
 
         @Positive
-        private int categoryRatio;
+        private double categoryRatio;
 
         private List<@Valid ProductDto> productList;
     }
@@ -48,7 +48,7 @@ public class CreateStrategyRequestDto {
         private String name;
 
         @Positive
-        private int ratio;
+        private double ratio;
 
         @NotBlank
         private String themeName;
